+++ src/main/java/com/modu/app/prj/pay/controller/KakaoPayController.java
@@ -33,17 +33,6 @@ public class KakaoPayController {
 		return payService.kakaoPayReady(payVO);
 	}
 	
-	// 결제성공
-	@GetMapping("/success")
-	public String successPay(@RequestParam("pg_token") String pgToken, Model model){
-		PayVO kakaoApprove = payService.approveResponse(pgToken);
-		// 성공시 db에 insert
-		payService.insertPay(kakaoApprove);
-		model.addAttribute("info", kakaoApprove);
-		// 결제완료페이지로
-		return "prj/success";
-	}
-	
 	// 결제 진행 중 취소
 	@GetMapping("/cancel")
 	public void cancel() {
