+++ src/main/resources/templates/prj/success.html
@@ -8,9 +8,9 @@
 	
 	<link rel="preconnect" href="https://fonts.gstatic.com">
 	<link href="https://fonts.googleapis.com/css2?family=Nunito:wght@300;400;600;700;800&display=swap" rel="stylesheet">
-	<link rel="stylesheet" href="assets/css/bootstrap.css">
-	<link rel="stylesheet" href="assets/vendors/bootstrap-icons/bootstrap-icons.css">
-	<link rel="stylesheet" href="assets/css/app.css">
+	<link rel="stylesheet" href="../assets/css/bootstrap.css">
+	<link rel="stylesheet" href="../assets/vendors/bootstrap-icons/bootstrap-icons.css">
+	<link rel="stylesheet" href="../assets/css/app.css">
     <style>
         .pay-card-body{
             text-align: left;
@@ -35,14 +35,16 @@
                 </h4>
             </div>
             <div class="card-body mx-auto pay-card-body">
-            [[${info}]]
+            [[${info.amount.total}]]
+            [[${info.approved_at}]]
+            [[${#strings.substring(info.approved_at, 0, 9)}]]
                     <p>결제일시 : <span>2023-08-03T12:42:05</span></p>
                     <p>결제금액 : <span>4800원</span></p>
                     <p>프로젝트 : <span>프로젝트 이름</span></p>
                     <p>기 간:  <span>2023-08-05 ~ 2023-09-04</span></p>
             </div>
             <div class="card-footer">
-                <button type="button" class="btn btn-light-primary mx-1" onclick="location.href='prjList'">
+                <button type="button" class="btn btn-light-primary mx-1" onclick="location.href='/modu/prjList'">
                     <i class="bx bx-x d-block d-sm-none"></i>
                     <span class="d-none d-sm-block">프로젝트 리스트로</span>
                 </button>
